rebase fixes AND move all redis tests to same place
