+++ crate/server/Cargo.toml
@@ -55,7 +55,13 @@ redis = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-sqlx = { version = "0.7.0-alpha.3", default-features = false, features = ["json", "runtime-tokio-native-tls", "mysql", "postgres", "sqlite"] }
+sqlx = { version = "0.7.0-alpha.3", default-features = false, features = [
+    "json",
+    "runtime-tokio-native-tls",
+    "mysql",
+    "postgres",
+    "sqlite",
+] }
 thiserror = { workspace = true }
 time = { workspace = true, features = ["local-offset", "formatting"] }
 tiny-keccak = { workspace = true }
@@ -69,6 +75,7 @@ zeroize = { workspace = true }
 [dev-dependencies]
 actix-http = "3.3"
 actix-rt = "2.8"
+serial_test = "2"
 
 [build-dependencies]
 time = { workspace = true, features = ["local-offset", "formatting"] }
