+++ crate/server/src/database/redis/mod.rs
@@ -3,5 +3,5 @@ mod redis_with_findex;
 
 // pub(crate) use redis_transaction_async::transaction_async;
 pub use redis_with_findex::{RedisWithFindex, REDIS_WITH_FINDEX_MASTER_KEY_LENGTH};
-mod objects_db;
-mod permissions;
+pub(crate) mod objects_db;
+pub(crate) mod permissions;
