+++ crate/server/src/database/redis/objects_db.rs
@@ -179,75 +179,3 @@ impl RemovedLocationsFinder for ObjectsDB {
         Ok(HashSet::new())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashSet;
-
-    use cloudproof::reexport::crypto_core::{
-        reexport::rand_core::{RngCore, SeedableRng},
-        CsRng,
-    };
-    use cosmian_kmip::kmip::kmip_types::{CryptographicAlgorithm, StateEnumeration};
-    use cosmian_kms_utils::crypto::symmetric::create_symmetric_key;
-    use redis::aio::ConnectionManager;
-    use serial_test::serial;
-    use tracing::trace;
-
-    use crate::{
-        database::redis::objects_db::{ObjectsDB, RedisDbObject},
-        log_utils::log_init,
-        result::KResult,
-    };
-
-    const REDIS_URL: &str = "redis://localhost:6379";
-
-    #[actix_web::test]
-    #[serial]
-    pub async fn test_objects_db() -> KResult<()> {
-        log_init("test_objects_db=trace");
-        trace!("test_objects_db");
-
-        let client = redis::Client::open(REDIS_URL)?;
-        let mgr = ConnectionManager::new(client).await?;
-
-        let o_db = ObjectsDB::new(mgr.clone()).await?;
-
-        // single upsert - get - delete
-        let uid = "test_objects_db";
-
-        let mut rng = CsRng::from_entropy();
-        let mut symmetric_key = vec![0; 32];
-        rng.fill_bytes(&mut symmetric_key);
-        let object = create_symmetric_key(&symmetric_key, CryptographicAlgorithm::AES);
-
-        // clean up
-        o_db.clear_all().await?;
-
-        // check that the object is not there
-        assert!(o_db.object_get(uid).await.is_err());
-
-        o_db.object_upsert(
-            uid,
-            &RedisDbObject::new(
-                object.clone(),
-                "owner".to_string(),
-                StateEnumeration::Active,
-                HashSet::new(),
-            ),
-        )
-        .await?;
-        let redis_db_object = o_db.object_get(uid).await?;
-        assert_eq!(
-            object.key_block()?.key_bytes()?,
-            redis_db_object.object.key_block()?.key_bytes()?
-        );
-        assert_eq!(redis_db_object.owner, "owner");
-        assert_eq!(redis_db_object.state, StateEnumeration::Active);
-
-        o_db.object_delete(uid).await?;
-        assert!(o_db.object_get(uid).await.is_err());
-
-        Ok(())
-    }
-}
