+++ crate/server/src/database/redis/permissions.rs
@@ -22,7 +22,7 @@ use crate::{error::KmsError, result::KResult};
 /// the database for all permissions for a given object or user because
 /// there is no convenient access to the callback for a search
 #[derive(Clone, Eq, PartialEq, Debug, Hash)]
-struct Triple {
+pub(crate) struct Triple {
     obj_uid: String,
     user_id: String,
     permission: ObjectOperationType,
@@ -318,388 +318,5 @@ impl RemovedLocationsFinder for PermissionsDB {
         _locations: HashSet<Location>,
     ) -> Result<HashSet<Location>, FindexRedisError> {
         Ok(HashSet::new())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{collections::HashSet, sync::Arc};
-
-    use async_trait::async_trait;
-    use cloudproof::reexport::{
-        crypto_core::{
-            reexport::rand_core::SeedableRng, CsRng, RandomFixedSizeCBytes, SymmetricKey,
-        },
-        findex::{
-            implementations::redis::{FindexRedis, FindexRedisError, RemovedLocationsFinder},
-            Location,
-        },
-    };
-    use cosmian_kms_utils::access::ObjectOperationType;
-    use redis::aio::ConnectionManager;
-    use serial_test::serial;
-
-    use crate::{database::redis::permissions::PermissionsDB, result::KResult};
-
-    const REDIS_URL: &str = "redis://localhost:6379";
-
-    struct DummyDB {}
-    #[async_trait]
-    impl RemovedLocationsFinder for DummyDB {
-        async fn find_removed_locations(
-            &self,
-            _locations: HashSet<Location>,
-        ) -> Result<HashSet<Location>, FindexRedisError> {
-            Ok(HashSet::new())
-        }
-    }
-
-    async fn clear_all(mgr: &mut ConnectionManager) -> KResult<()> {
-        redis::cmd("FLUSHDB").query_async(mgr).await?;
-        Ok(())
-    }
-
-    #[actix_web::test]
-    #[serial]
-    pub async fn test_permissions_db() -> KResult<()> {
-        // generate the findex key
-        let mut rng = CsRng::from_entropy();
-        let findex_key = SymmetricKey::new(&mut rng);
-
-        // the findex label
-        let label = b"label";
-
-        let client = redis::Client::open(REDIS_URL)?;
-        let mut mgr = ConnectionManager::new(client).await?;
-        // clear the DB
-        clear_all(&mut mgr).await?;
-        // create the findex
-        let findex =
-            Arc::new(FindexRedis::connect_with_manager(mgr.clone(), Arc::new(DummyDB {})).await?);
-        let permissions_db = PermissionsDB::new(findex, label).await?;
-
-        // let us add the permission Encrypt on object O1 for user U1
-        permissions_db
-            .add(&findex_key, "O1", "U1", ObjectOperationType::Encrypt)
-            .await?;
-
-        // verify that the permission is present
-        let permissions = permissions_db.get(&findex_key, "O1", "U1").await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains(&ObjectOperationType::Encrypt));
-
-        // find the permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("U1"));
-        assert_eq!(
-            permissions.get("U1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // add the permission Decrypt to user U1 for object O1
-        permissions_db
-            .add(&findex_key, "O1", "U1", ObjectOperationType::Decrypt)
-            .await?;
-
-        // assert the permission is present
-        let permissions = permissions_db.get(&findex_key, "O1", "U1").await?;
-        assert_eq!(permissions.len(), 2);
-        assert!(permissions.contains(&ObjectOperationType::Encrypt));
-        assert!(permissions.contains(&ObjectOperationType::Decrypt));
-
-        // find the permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt, ObjectOperationType::Decrypt])
-        );
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("U1"));
-        assert_eq!(
-            permissions.get("U1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt, ObjectOperationType::Decrypt])
-        );
-
-        // the situation now is that we have
-        // O1 -> U1 -> Encrypt, Decrypt
-
-        // let us add the permission Encrypt on object O1 for user U2
-        permissions_db
-            .add(&findex_key, "O1", "U2", ObjectOperationType::Encrypt)
-            .await?;
-        // assert the permission is present
-        let permissions = permissions_db.get(&findex_key, "O1", "U2").await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains(&ObjectOperationType::Encrypt));
-
-        // find the permissions for user U2
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U2")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 2);
-        assert!(permissions.contains_key("U1"));
-        assert_eq!(
-            permissions.get("U1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt, ObjectOperationType::Decrypt])
-        );
-        assert!(permissions.contains_key("U2"));
-        assert_eq!(
-            permissions.get("U2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // the situation now is that we have
-        // O1 -> U1 -> Encrypt, Decrypt
-        // O1 -> U2 -> Encrypt
-
-        // let us add the permission Encrypt on object O2 for user U2
-        permissions_db
-            .add(&findex_key, "O2", "U2", ObjectOperationType::Encrypt)
-            .await?;
-        // assert the permission is present
-        let permissions = permissions_db.get(&findex_key, "O2", "U2").await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains(&ObjectOperationType::Encrypt));
-
-        // find the permissions for user U2
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U2")
-            .await?;
-        assert_eq!(permissions.len(), 2);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-        assert!(permissions.contains_key("O2"));
-        assert_eq!(
-            permissions.get("O2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        //find the permission for the object O2
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O2")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("U2"));
-        assert_eq!(
-            permissions.get("U2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // the situation now is that we have
-        // O1 -> U1 -> Encrypt, Decrypt
-        // O1 -> U2 -> Encrypt
-        // O2 -> U2 -> Encrypt
-
-        // let us remove the permission Decrypt on object O1 for user U1
-        permissions_db
-            .remove(&findex_key, "O1", "U1", ObjectOperationType::Decrypt)
-            .await?;
-        // assert the permission Encrypt is present and Decrypt is not
-        let permissions = permissions_db.get(&findex_key, "O1", "U1").await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains(&ObjectOperationType::Encrypt));
-
-        // find the permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 2);
-        assert!(permissions.contains_key("U1"));
-        assert_eq!(
-            permissions.get("U1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-        assert!(permissions.contains_key("U2"));
-        assert_eq!(
-            permissions.get("U2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // let us remove the permission Encrypt on object O1 for user U1
-        permissions_db
-            .remove(&findex_key, "O1", "U1", ObjectOperationType::Encrypt)
-            .await?;
-        // assert the permission is not present
-        let permissions = permissions_db.get(&findex_key, "O1", "U1").await?;
-        assert_eq!(permissions.len(), 0);
-
-        // find the permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("U2"));
-        assert_eq!(
-            permissions.get("U2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // let us remove the permission Encrypt on object O1 for user U2
-        permissions_db
-            .remove(&findex_key, "O1", "U2", ObjectOperationType::Encrypt)
-            .await?;
-        // assert the permission is not present
-        let permissions = permissions_db.get(&findex_key, "O1", "U2").await?;
-        assert_eq!(permissions.len(), 0);
-
-        // find the permissions for user U2
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U2")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O2"));
-        assert_eq!(
-            permissions.get("O2").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        //find the permission for the object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        Ok(())
-    }
-
-    #[actix_web::test]
-    #[serial]
-    pub async fn test_corner_case() -> KResult<()> {
-        // generate the findex key
-        let mut rng = CsRng::from_entropy();
-        let findex_key = SymmetricKey::new(&mut rng);
-
-        // the findex label
-        let label = b"label";
-
-        let client = redis::Client::open(REDIS_URL)?;
-        let mut mgr = ConnectionManager::new(client).await?;
-        // clear the DB
-        clear_all(&mut mgr).await?;
-        // create the findex
-        let findex =
-            Arc::new(FindexRedis::connect_with_manager(mgr.clone(), Arc::new(DummyDB {})).await?);
-        let permissions_db = PermissionsDB::new(findex, label).await?;
-
-        // remove a permission that does not exist
-        permissions_db
-            .remove(&findex_key, "O1", "U1", ObjectOperationType::Encrypt)
-            .await?;
-
-        // test that it does not exist
-        let permissions = permissions_db.get(&findex_key, "O1", "U1").await?;
-        assert_eq!(permissions.len(), 0);
-
-        // test there are no permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        // test there are no permissions for object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        //add the permission Encrypt on object O1 for user U1
-        permissions_db
-            .add(&findex_key, "O1", "U1", ObjectOperationType::Encrypt)
-            .await?;
-
-        // test there is one permission for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("O1"));
-        assert_eq!(
-            permissions.get("O1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // test there is one permission for object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 1);
-        assert!(permissions.contains_key("U1"));
-        assert_eq!(
-            permissions.get("U1").unwrap(),
-            &HashSet::from([ObjectOperationType::Encrypt])
-        );
-
-        // remove the permission again
-        permissions_db
-            .remove(&findex_key, "O1", "U1", ObjectOperationType::Encrypt)
-            .await?;
-
-        // test there are no permissions for user U1
-        let permissions = permissions_db
-            .list_user_permissions(&findex_key, "U1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        // test there are no permissions for object O1
-        let permissions = permissions_db
-            .list_object_permissions(&findex_key, "O1")
-            .await?;
-        assert_eq!(permissions.len(), 0);
-
-        Ok(())
     }
 }
