+++ crate/server/src/tests/curve_25519_tests.rs
@@ -8,7 +8,6 @@ use cosmian_kmip::kmip::{
     },
 };
 use cosmian_kms_utils::crypto::curve_25519::{
-    encryption_decryption::X25519_PUBLIC_KEY_LENGTH,
     kmip_requests::{create_key_pair_request, get_private_key_request, get_public_key_request},
     operation::{self, to_curve_25519_256_public_key},
     X25519_PUBLIC_KEY_LENGTH,
