+++ crate/server/src/tests/kmip_server_tests.rs
@@ -10,7 +10,6 @@ use cosmian_kmip::kmip::{
     },
 };
 use cosmian_kms_utils::crypto::curve_25519::{
-    encryption_decryption::X25519_PUBLIC_KEY_LENGTH,
     kmip_requests::{create_key_pair_request, get_private_key_request, get_public_key_request},
     operation::{to_curve_25519_256_public_key, Q_LENGTH_BITS},
     X25519_PUBLIC_KEY_LENGTH,
